move network metadata into a hook
