+++ src/components/atoms/NetworkName.tsx
@@ -1,30 +1,11 @@
-import { graphql, useStaticQuery } from 'gatsby'
 import React, { ReactElement } from 'react'
 import { ReactComponent as EthIcon } from '../../images/eth.svg'
 import { ReactComponent as PolygonIcon } from '../../images/polygon.svg'
 import { ReactComponent as MoonbeamIcon } from '../../images/moonbeam.svg'
 import { ReactComponent as BscIcon } from '../../images/bsc.svg'
-import {
-  EthereumListsChain,
-  getNetworkDataById,
-  getNetworkDisplayName
-} from '../../utils/web3'
+import { getNetworkDataById, getNetworkDisplayName } from '../../utils/web3'
 import styles from './NetworkName.module.css'
-
-const networksQuery = graphql`
-  query {
-    allNetworksMetadataJson {
-      edges {
-        node {
-          chain
-          network
-          networkId
-          chainId
-        }
-      }
-    }
-  }
-`
+import useNetworkMetadata from '../../hooks/useNetworkMetadata'
 
 export function NetworkIcon({ name }: { name: string }): ReactElement {
   const IconMapped = name.includes('ETH')
@@ -49,9 +30,7 @@ export default function NetworkName({
   minimal?: boolean
   className?: string
 }): ReactElement {
-  const data = useStaticQuery(networksQuery)
-  const networksList: { node: EthereumListsChain }[] =
-    data.allNetworksMetadataJson.edges
+  const { networksList } = useNetworkMetadata()
   const networkData = getNetworkDataById(networksList, networkId)
   const networkName = getNetworkDisplayName(networkData, networkId)
